% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmr.weakiv.R
\name{mvmr.weakiv}
\alias{mvmr.weakiv}
\title{Weak instrument-robust inference for multivariable Mendelian randomization with two-sample summary data}
\usage{
mvmr.weakiv(
  bx,
  by,
  sx,
  sy,
  nx,
  ny,
  ld = NULL,
  cor.x = NULL,
  alpha = NULL,
  gamma.min = NULL,
  max.search = NULL,
  len.search = NULL,
  robust = FALSE,
  kappa.limit = NULL,
  seed = 100
)
}
\arguments{
\item{bx}{J x K matrix of genetic variant associations with K exposures (number of variants J cannot be less than number of exposures K)}

\item{by}{J vector of genetic variant associations with outcome}

\item{sx}{J x K matrix of corresponding standard errors from variant-exposure regressions}

\item{sy}{J vector of corresponding standard errors from variant-outcome regressions}

\item{nx}{sample size for genetic variant-exposure regressions; the exposure associations are assumed to be measured from the same sample}

\item{ny}{sample size for genetic variant-outcome regressions}

\item{ld}{(optional) J x J genetic variant correlation matrix; if not supplied, the default is to assume uncorrelated variants}

\item{cor.x}{(optional) K x K matrix exposure correlation matrix; if not supplied, the default is to assume uncorrelated exposures}

\item{alpha}{(optional) nominal significance level; 1-alpha is the nominal coverage probability of confidence sets}

\item{gamma.min}{(optional) minimum allowable coverage distortion for non-robust inference; if not supplied, the default is 0.05}

\item{max.search}{(optional) c(-max.search, max.search) are the limits for the search if a positive scalar is provided, or else a vector with two numbers may be provided so that c(max.search1,max.search2) are the limits for each exposure. If not supplied, default is max.search = 2}

\item{len.search}{(optional) number of potential parameter values for each exposure to be tested to form confidence sets. The runtime substantially increases with larger len.search, and exponentially so with the number of exposures. If not supplied, default is len.search=100}

\item{robust}{(optional) logical \code{TRUE}/\code{FALSE}. If \code{TRUE}, the Kleibergen-OH confidence set is computed, but runtime may take a lot longer. The default is \code{FALSE}}

\item{kappa.limit}{(optional) c(-kappa.limit, kappa.limit) are the limits for the overdispersion parameter estimate. Only relevant if \code{robust} is \code{TRUE} and for the Kleibergen-OH confidence set. If not supplied, default is kappa.limit=250}

\item{seed}{(optional) the seed for random sampling involved in computing critical values for the Andrews robust confidence set (default value is 100)}
}
\value{
Output is a list containing:
\itemize{
\item \code{condFstat} \cr
K vector of conditional F statistics (one for each exposure)
\item \code{gmm_est} \cr
K vector of causal effect estimates using GMM estimation
\item \code{cs_n} \cr
non-robust confidence set based on inverting Wald test statistics; each column is a K vector of exposure effect values not rejected the test
\item \code{cs_ar} \cr
robust confidence set based on inverting Anderson-Rubin test statistics; each column is a K vector of exposure effect values not rejected by the test
\item \code{cs_k} \cr
robust confidence set based on inverting Kleibergen test statistics; each column is a K vector of exposure effect values not rejected by the test
\item \code{cs_koh} \cr
robust confidence set based on inverting Kleibergen-OH test statistics that account for overdispersion heterogeneity; each column is a K vector of exposure effect values not rejected by the test. Reported only if the input \code{robust} is \code{TRUE}
\item \code{cs_r} \cr
robust confidence set based on inverting Andrews linear combination test statistics; each column is a K vector of exposure effect values not rejected by the test
\item \code{cutoff} \cr
maximum coverage distortion for non-robust Wald-based confidence set \code{cs_n}; minimum value is gamma.min
\item \code{kappa} \cr
the average estimated overdispersion heterogeneity parameter over non-rejected values of exposure effects. Reported only if the input \code{robust} is \code{TRUE}
}
}
\description{
Compute weak instrument-robust confidence sets in multivariable Mendelian randomization with two-sample summary data.
}
\details{
Weak instrument-robust confidence sets for multivariable Mendelian randomization with two-sample summary data. \cr
\cr
The calculated distortion cutoff is the estimated maximum coverage distortion relating to only the non-robust confidence set (based on inverting Wald test statistics). \cr
\cr
The runtime for computing confidence sets substantially increases with larger values of len.search, especially with large numbers of instruments, and the runtime is exponentially longer with every additional exposure included in the model.
}
\examples{
res <- mvmr.weakiv(bx=cbind(dat$beta.map,dat$beta.pp),by=dat$beta.stroke,sx=cbind(dat$se.map,dat$se.pp),sy=dat$se.stroke,nx=dat$nx,ny=dat$ny,cor.x=dat$cor.x,max.search=0.3,len.search=80)
# this may take a few minutes to run

# note that the 95\% Anderson-Rubin confidence set res$cs_ar is empty. The Anderson-Rubin test doubles as an overidentification test, so the confidence set being empty may suggest excessive heterogeneity in genetic variant-outcome associations in the model

# we can plot the 95\% non-robust (Wald) and robust (Andrews) confidence sets
plot(res$cs_n[1,],res$cs_n[2,],pch=16,main="Non-robust (Wald) confidence set", xlab = "MAP effect", ylab = "PP effect", xlim=c(-0.3,0.3), ylim=c(-0.3,0.3))
grid();abline(v=0,lty=2);abline(h=0,lty=2)
plot(res$cs_r[1,],res$cs_r[2,],pch=16,main="Robust (Andrews) confidence set", xlab = "MAP effect", ylab = "PP effect", xlim=c(-0.3,0.3), ylim=c(-0.3,0.3))
grid();abline(v=0,lty=2);abline(h=0,lty=2)

# the distortion cutoff res$cutoff is around 15\%, which means we may value the evidence from the non-robust set if we are willing to accept that the non-robust set may have coverage of at least 80\%

}
\references{
Weak instruments in multivariable Mendelian randomization: methods and practice. Preprint.
}
\author{
Ashish Patel and Stephen Burgess
}
